TouchEvent/HoverEvent/WheelEvent: complete the public creation API
- TouchEvent/HoverEvent: add New(time), AddPoint() and SetTime()
- Promote DevelWheelEvent::New to WheelEvent::New
- Add Integration event constructors that take the public handle
- TouchEvent/HoverEvent time → uint32_t
- Remove the devel and integration APIs this replaces

+++ dali/internal/event/events/hover-event-impl.cpp
@@ -28,7 +28,7 @@ HoverEvent::HoverEvent()
 {
 }
 
-HoverEvent::HoverEvent(unsigned long time)
+HoverEvent::HoverEvent(uint32_t time)
 : mPoints(),
   mTime(time)
 {
@@ -42,7 +42,7 @@ HoverEventPtr HoverEvent::Clone(const HoverEvent& rhs)
   return hoverEvent;
 }
 
-unsigned long HoverEvent::GetTime() const
+uint32_t HoverEvent::GetTime() const
 {
   return mTime;
 }
