+++ dali/internal/event/events/touch-event-impl.cpp
@@ -168,7 +168,7 @@ void TouchEvent::AddPoint(const Integration::Point& point)
   mPoints.push_back(point);
 }
 
-void TouchEvent::SetTime(uint64_t time)
+void TouchEvent::SetTime(uint32_t time)
 {
   mTime = time;
 }
